Move table migration ownership to table_migrate module
